Moved dependencies to deps.ts
