+++ acciones/frase.ts
@@ -1,4 +1,4 @@
-import Bot from "../types/bot.type.ts";
+import { Bot } from "../deps.ts";
 import getBotonesFrases from "./getBotonesFrases.ts";
 import { frasesDB } from "../db/collections/collections.ts";
