+++ deps.ts
@@ -1,0 +1,4 @@
+export { sleep } from "sleep";
+export { default as loki } from "lokijs";
+export { lodash as _, lodash } from "lodash";
+export { Bot, InlineKeyboard } from "grammy/mod.ts";
