+++ import_map.json
@@ -2,13 +2,8 @@
   "imports": {
     "lokijs": "https://esm.sh/lokijs@1.5.12",
     "grammy/": "https://deno.land/x/grammy@v1.14.1/",
-    "axiod": "https://deno.land/x/axiod@0.26.2/mod.ts",
     "sleep": "https://deno.land/x/sleep@v1.2.1/mod.ts",
-    "join": "https://deno.land/std@0.179.0/path/mod.ts",
-    "croner": "https://deno.land/x/croner/dist/croner.js",
     "dotenv": "https://deno.land/x/dotenv@v3.2.0/load.ts",
-    "lodash": "https://deno.land/x/deno_ts_lodash@0.0.1/mod.ts",
-    "escapeHtml": "https://deno.land/x/escape_html@1.0.0/mod.ts",
-    "duplicatedFilesCleanerIncognito": "https://deno.land/x/duplicated_files_cleaner_incognito@1.1.8/mod.ts"
+    "lodash": "https://deno.land/x/deno_ts_lodash@0.0.1/mod.ts"
   }
 }
