--- types/bot.type.ts
@@ -1,2 +1,0 @@
-import { Bot } from "grammy/mod.ts";
-export default Bot;
